requirements.txt
pytorch>=1.8.0
